work on items in the shop
connected SOItemDataObjects to the shop, and time cost + currency cost is displayed in the shopping cart window. however it is not displaying correctly sometimes -> when you drag the object just a bit away from the shopping cart it resets the cost and it's still in the shopping cart. will work on that tomorrow as it shouldn't be too difficult to fix.
will start on the inventory after that.
also the structuring is a mess and i will deal with it tomorrow as well to make it more pleasing and more easily understandable tomorrow </3